+++ packages/charts/src/lib/index.js
@@ -4,10 +4,6 @@ export { default as Footer } from './chartContainer/Footer.svelte';
 export { default as SubTitle } from './chartContainer/SubTitle.svelte';
 export { default as Title } from './chartContainer/Title.svelte';
 
-export { default as ScatterPlot } from './scatterPlot/ScatterPlot.svelte';
-
 export { default as ObservablePlot } from './observablePlot/ObservablePlot.svelte';
 
-export { default as AxisX } from './shared/AxisX.svelte';
-export { default as AxisY } from './shared/AxisY.svelte';
-export { default as Scatter } from './shared/Scatter.svelte';
+export * from './observablePlotFragments/observablePlotFragments';