+++ packages/ui/src/lib/button/Button.stories.svelte
@@ -72,19 +72,21 @@
 
 <Story name="Outline" source args={{ variant: 'outline' }} />
 
-<Story name="Text" source args={{ variant: 'text' }} />
+<Story name="Text" source args={{ variant: 'text', emphasis: 'secondary' }} />
 
 <Story name="Square" args={{ variant: 'square' }}>
-	<Button variant="square" title="Down" size="sm">
-		<Icon src={ArrowDownCircle} theme="solid" class="w-6 h-6" aria-hidden="true" />
-	</Button>
-	<Button variant="square" title="Down" size="md">
-		<Icon src={ArrowDownCircle} theme="solid" class="w-6 h-6" aria-hidden="true" />
-	</Button>
-	<Button variant="square" title="Down" size="lg">
-		<Icon src={ArrowDownCircle} theme="solid" class="w-8 h-8" aria-hidden="true" />
-		Down
-	</Button>
+	<div class="space-y-2">
+		<Button variant="square" title="Down" size="sm">
+			<Icon src={ArrowDownCircle} theme="mini" class="w-5 h-5" aria-hidden="true" />
+		</Button>
+		<Button variant="square" title="Down" size="md">
+			<Icon src={ArrowDownCircle} theme="solid" class="w-6 h-6" aria-hidden="true" />
+		</Button>
+		<Button variant="square" title="Down" size="lg">
+			<Icon src={ArrowDownCircle} theme="solid" class="w-8 h-8" aria-hidden="true" />
+			Down
+		</Button>
+	</div>
 </Story>
 
 <Story name="Disabled">
@@ -96,11 +98,17 @@
 		<Button size="sm">sm</Button>
 		<Button size="md">md</Button>
 		<Button size="lg">lg</Button>
+		<Button variant="outline" title="Down" emphasis="primary">Down</Button>
+
+		<Button variant="outline" title="Down" emphasis="secondary">Down</Button>
+		<Button emphasis="secondary">Down</Button>
 	</div>
 </Story>
 
 <Story name="With class prop">
-	<Button class="text-core-darkPink-500 border-core-pink-700 bg-yellow-500">
+	<Button
+		class="text-color-palette-pink-100 border-2 border-color-palette-pink-700 bg-color-palette-yellow-500"
+	>
 		Custom classes applied
 	</Button>
 </Story>
@@ -117,11 +125,13 @@
 <Story name="With Icon">
 	<Button>
 		Download as PNG
-		<Icon src={Camera} theme="solid" class="ml-2 w-6 h-6" aria-hidden="true" />
+		<Icon src={Camera} theme="mini" class="ml-2 w-5 h-5" aria-hidden="true" />
 	</Button>
 </Story>
 
 <Story name="With Link">
-	<Button href="#">Link</Button>
-	<Button disabled href="#">Link</Button>
+	<div class="space-y-2">
+		<Button href="#">Link</Button>
+		<Button disabled href="#">Link</Button>
+	</div>
 </Story>