+++ packages/ui/src/lib/input/Input.stories.svelte
@@ -2,6 +2,34 @@
 	import { Story, Template } from '@storybook/addon-svelte-csf';
 	import Input from './Input.svelte';
 
+	const newStringArg = (defaultValue = '') => ({
+		control: { type: 'text' },
+		table: {
+			defaultValue: { summary: defaultValue },
+			type: { summary: 'string' }
+		}
+	});
+
+	const newSelectedStringArg = (options, defaultValue = '') => ({
+		control: { type: 'select' },
+		options: options,
+		table: {
+			defaultValue: { summary: defaultValue },
+			type: { summary: 'string' }
+		}
+	});
+
+	const newFunctionArg = (detail = '') => ({
+		type: 'function',
+		control: 'none',
+		table: {
+			type: {
+				summary: 'function',
+				detail: detail
+			}
+		}
+	});
+
 	export const meta = {
 		title: 'Ui/Input',
 		component: Input,
@@ -12,10 +40,47 @@
 					disable: true
 				}
 			},
+			id: newStringArg('// random ID'),
+			name: newStringArg("// 'id' prop"),
+			type: newSelectedStringArg(
+				[
+					'text',
+					'textarea',
+					'number',
+					'email',
+					'password',
+					'search',
+					'tel',
+					'url',
+					'time',
+					'date',
+					'datetime-local',
+					'file',
+					'range'
+				],
+				'text'
+			),
+			inputmode: newSelectedStringArg(
+				[
+					'none',
+					'text',
+					'decimal',
+					'numeric',
+					'tel',
+					'search',
+					'email',
+					'url',
+					'search',
+					'tel',
+					'url'
+				],
+				'text'
+			),
 			descriptionAlignment: {
-				options: ['left', 'right'],
-				control: { type: 'select' }
-			}
+				control: { type: 'select' },
+				options: ['left', 'right']
+			},
+			format: newFunctionArg('(value: string) => string')
 		}
 	};
 </script>
@@ -28,7 +93,9 @@
 
 <Template let:args>
 	<div class="w-96">
-		<Input {...args} />
+		{#key args}
+			<Input {...args} id="1" />
+		{/key}
 	</div>
 </Template>
 
@@ -36,7 +103,7 @@
 
 <Story name="Binding on value">
 	<div class="w-96 flex flex-col gap-4">
-		<Input name="bind-value-input" bind:value placeholder="Type here..." />
+		<Input name="bind-value-input" bind:value placeholder="Type here..." id="1" />
 		<div class="h-4">
 			<span class="font-medium dark:text-white">Value:</span>
 			{value}
@@ -47,13 +114,19 @@
 <Story name="With Label">
 	<div class="w-96 flex flex-col gap-4">
 		<Input name="labelled-input-required" label="Label" />
-		<Input name="labelled-input-optional" label="Label" optional />
+		<Input name="labelled-input-optional" label="Label" />
 	</div>
 </Story>
 
 <Story name="With tooltip - default hintLabel">
 	<div class="w-96">
-		<Input label="Tooltip" name="tooltip-input" hintLabel="optional hint label" />
+		<Input label="Tooltip" name="tooltip-input" />
+	</div>
+</Story>
+
+<Story name="With placeholder">
+	<div class="w-96">
+		<Input label="Placeholder" name="placeholder-input" placeholder="This is placeholder text" />
 	</div>
 </Story>
 
@@ -64,6 +137,7 @@
 			name="tooltip-input"
 			hint="Contextual help text"
 			hintLabel="optional hint label"
+			id="1"
 		/>
 	</div>
 </Story>
@@ -74,19 +148,21 @@
 			label="Label"
 			name="left-aligned-description-input"
 			description="Left aligned descriptive text (default)."
+			id="1"
 		/>
 		<Input
 			label="Label"
 			name="right-aligned-description-input"
 			description="Right aligned descriptive text."
 			descriptionAlignment="right"
+			id="1"
 		/>
 	</div>
 </Story>
 
 <Story name="With error">
 	<div class="w-96">
-		<Input label="Description" name="description-input" error="Error text" />
+		<Input label="Description" name="description-input" error="Error text" id="1" />
 	</div>
 </Story>
 
@@ -98,13 +174,14 @@
 			inputmode="decimal"
 			name="input-mode-input"
 			hint="Using 'decimal' input mode here rather than 'numeric'"
+			id="1"
 		/>
 	</div>
 </Story>
 
 <Story name="Disabled">
 	<div class="w-96">
-		<Input label="Disabled" name="disabled-input" disabled />
+		<Input label="Disabled" name="disabled-input" disabled id="1" />
 	</div>
 </Story>
 
@@ -114,6 +191,7 @@
 			label="Type a number"
 			description="It will be rounded to 2 d.p. when the input loses focus"
 			format={round}
+			id="1"
 		/>
 	</div>
 </Story>
@@ -126,31 +204,42 @@
 			name="textarea-input"
 			rows="5"
 			hint="Three rows by default"
+			id="1"
 		/>
 		<Input
 			type="number"
 			label="Number"
 			name="number-input"
 			hint="Brings up the numeric keypad on touch screens"
+			id="1"
 		/>
-		<Input type="password" label="Password" name="password-input" hint="No formatting by default" />
+		<Input
+			type="password"
+			label="Password"
+			name="password-input"
+			hint="No formatting by default"
+			id="1"
+		/>
 		<Input
 			type="url"
 			label="URL"
 			name="url-input"
 			hint="Brings up the URL keypad on touch screens"
+			id="1"
 		/>
 		<Input
 			type="tel"
 			label="Phone"
 			name="tel-input"
 			hint="Brings up the telephone keypad on touch screens"
+			id="1"
 		/>
 		<Input
 			type="email"
 			label="Email"
 			name="email-input"
 			hint="Brings up the email keypad on touch screens"
+			id="1"
 		/>
 		<Input type="time" label="Time" name="time-input" />
 		<Input type="date" label="Date" name="date-input" />