+++ packages/ui/src/lib/sidebar/elements/sidebarTabs/SidebarTabLabel.svelte
@@ -1,3 +1,33 @@
+<script lang="ts" context="module">
+	import { get } from 'svelte/store';
+
+	/**
+	 * The default function called when a user clicks on a `TabLabel`.
+	 */
+	export const handleSelection = (
+		sidebarIsOpen: Writable<boolean>,
+		sidebarAlwaysOpen: Writable<'true' | 'false'>,
+		selectedValue: Writable<string>,
+		tabId: string
+	) => {
+		if (get(sidebarIsOpen) === false) {
+			// if we're collapsed, clicking any tab label will open that tab
+			sidebarIsOpen.set(true);
+			selectedValue.set(tabId);
+		} else if (get(selectedValue) === tabId) {
+			// The sidebarAlways open context is provided by the app shell
+			if (get(sidebarAlwaysOpen) === 'false' || get(sidebarAlwaysOpen) === undefined) {
+				// if we're expanded, clicking the currently selected tab label triggers collapse
+				sidebarIsOpen.set(false);
+				selectedValue.set('');
+			}
+		} else {
+			// if we're expanded, clicking a different tab label switched tab
+			selectedValue.set(tabId);
+		}
+	};
+</script>
+
 <script lang="ts">
 	import { getContext } from 'svelte';
 	import type { Writable } from 'svelte/store';
@@ -19,35 +49,20 @@
 	// Context provided by wrapping component to force always open
 	const sidebarAlwaysOpen = getContext<Writable<'true' | 'false'>>('sidebarAlwaysOpen');
 
-	const handleSelection = () => {
-		if ($sidebarIsOpen === false) {
-			// if we're collapsed, clicking any tab label will open that tab
-			$sidebarIsOpen = true;
-			$selectedValue = tabId;
-		} else if ($selectedValue === tabId) {
-			// The sidebarAlways open context is provided by the app shell
-			if ($sidebarAlwaysOpen === 'false' || $sidebarAlwaysOpen === undefined) {
-				// if we're expanded, clicking the currently selected tab label triggers collapse
-				$sidebarIsOpen = false;
-				$selectedValue = '';
-			}
-		} else {
-			// if we're expanded, clicking a different tab label switched tab
-			$selectedValue = tabId;
-		}
-	};
+	/**
+	 * Function called when user clicks on a tab label.
+	 */
+	export let selectionHandler = handleSelection;
 
 	const keydownHandler = (ev: KeyboardEvent) => {
 		if (ev.key === 'Enter' || ev.key === ' ') {
-			handleSelection();
+			selectionHandler(sidebarIsOpen, sidebarAlwaysOpen, selectedValue, tabId);
 		}
 	};
 
-	const darkThemeClasses = 'dark:bg-core-grey-900 dark:hover:bg-core-blue-700 dark:text-white';
-
-	const lightThemeClasses = 'bg-core-grey-100 hover:bg-core-grey-200 text-core-grey-800';
-
-	const themeClasses = [darkThemeClasses, lightThemeClasses];
+	const themeClasses = [
+		'bg-color-action-background-secondary hover:bg-color-action-background-secondary-hover text-color-text-primary'
+	];
 
 	const orientationClasses = {
 		vertical:
@@ -59,13 +74,15 @@
 </script>
 
 <div
-	on:click={handleSelection}
+	on:click={() => selectionHandler(sidebarIsOpen, sidebarAlwaysOpen, selectedValue, tabId)}
 	on:keydown={keydownHandler}
 	tabindex="0"
 	role="tab"
 	class={classNames(
 		tabLabelClasses,
-		tabId === $selectedValue ? '!bg-core-blue-600 cursor-default text-white' : 'cursor-pointer'
+		tabId === $selectedValue
+			? '!bg-color-action-background-primary cursor-default !text-color-static-white hover:!bg-color-action-background-primary-hover'
+			: 'cursor-pointer'
 	)}
 >
 	<!-- Label and/or icon -->