+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,42 @@
 # ldn-viz-tools-ui
 
+## 13.0.0
+
+### Major Changes
+
+- b55dab9: CHANGED: moved `userThemeSelectionStore` and `currentThemeMode` from `theme` package to `ui` package
+
+## 12.0.0
+
+### Major Changes
+
+- bfad1da: CHANGED: Move `userPreference` from `utils` package to `ui` package
+
+### Patch Changes
+
+- Updated dependencies [bfad1da]
+  - @ldn-viz/utils@1.0.0
+
+## 11.0.1
+
+### Patch Changes
+
+- 69686cf: FIXED: adds missing `@ldn-viz/util` dependency to `package.json` of `maps`, `ui`, and `charts` packages.
+
+## 11.0.0
+
+### Major Changes
+
+- 9f1adf7: CHANGED: change how we handle color tokens
+
+### Minor Changes
+
+- 9f1adf7: ADDED: add `Theme` and `ThemeSwitcher` components
+
+### Patch Changes
+
+- 881aa44: fix text color of input when disabled
+
 ## 10.1.0
 
 ### Minor Changes