+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ldn-viz/utils
 
+## 1.0.0
+
+### Major Changes
+
+- bfad1da: CHANGED: Move `userPreference` from `utils` package to `ui` package
+
+## 0.2.0
+
+### Minor Changes
+
+- 9f1adf7: ADDED: add `mediaQueryStore`, `prefersDarkMode`, `prefersReducedMotion` functions
+
 ## 0.1.0
 
 ### Minor Changes